# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MAXMOLL - REST API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/stock-movements": {
            "get": {
                "tags": [
                    "Движение товаров"
                ],
                "summary": "Получить движение товаров с фильтром и пагинацией",
                "operationId": "6612cab7ac7d470c751b001f795a07ac",
                "parameters": [
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Filter by warehouse ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filter by product ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_id",
                        "in": "query",
                        "description": "Filter by order ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter movements created after this date (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter movements created before this date (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "customer",
                        "in": "query",
                        "description": "Filter by customer name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (1-100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "StockMovement": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 74
                                                            },
                                                            "product": {
                                                                "type": "string",
                                                                "example": "quis impedit debitis"
                                                            },
                                                            "warehouse": {
                                                                "type": "string",
                                                                "example": "Dinoton"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 4
                                                            },
                                                            "balance_after": {
                                                                "type": "integer",
                                                                "example": 111088
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-07-26 10:00:43"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "links": {
                                                        "properties": {
                                                            "self": {
                                                                "type": "string",
                                                                "example": "https://max/api/v1/orders/74"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "https://max/api/v1/stock-movements?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "https://max/api/v1/stock-movements?page=5"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "https://max/api/v1/stock-movements?page=2"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "https://max/api/v1/stock-movements"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 74
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid date format"
                                        },
                                        "errors": {
                                            "properties": {
                                                "date_from": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The date_from does not match the format Y-m-d."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/stock-movements/{stock_movement}": {
            "get": {
                "tags": [
                    "Движение товаров"
                ],
                "summary": "Получить движение по ID",
                "operationId": "7a483f1d9d9c8d18b86d27fbf6149d5f",
                "parameters": [
                    {
                        "name": "stock_movement",
                        "in": "path",
                        "description": "ID of the stock movement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock movement retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "StockMovement": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "product": {
                                                            "type": "string",
                                                            "example": "quia repudiandae perferendis"
                                                        },
                                                        "warehouse": {
                                                            "type": "string",
                                                            "example": "Dinoton"
                                                        },
                                                        "quantity": {
                                                            "type": "integer",
                                                            "example": 884
                                                        },
                                                        "balance_after": {
                                                            "type": "integer",
                                                            "example": 326
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-07-25 10:23:47"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "links": {
                                                    "properties": {
                                                        "self": {
                                                            "type": "string",
                                                            "example": "https://max/api/v1/orders/1"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "StockMovement №'1' retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Stock movement not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Stock movement not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}